PythonProject/requirements.txt
discord.py
python-dotenv
flask
flask-sqlalchemy
gunicorn
requests
psycopg2-binary